+++ app/code/local/PayU/Account/Model/Session.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- *	ver. 0.1.6.2
+ *	ver. 0.1.6.3
  *	PayU Session Model
  *
  *	@copyright  Copyright (c) 2011-2012 PayU