+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.6.3
+
+* Changed shopping process flow without authentication before summary
+* Fixed billing address for virtual order
+
 ## 0.1.6.2
 
 * Fixed shipping costs list for virtual order